[Tests] Move asset provider test classes
